Access program settings via specialized controller
This commit creates a seperate class for handling program settings.
This allows other classes to not worry about how to save settings.
They just provide the contents to save, and the new controller will
take care of actually saving them.  This also allows multiple
components to access the settings from one place.  The direction
things were going in before would have required all classes to
implement their logic for reading and writing, which would have
been a mess.

Much code has been removed.  The methods to get and set colors in
DataOptionsController have been moved to the new SettingsController
because they are related more to the settings.  The logic for reading
and writing settings in CustomizedCalendarDialog has been moved to
SettingsController.  The dialog now only provides objects to save.